(Staticman) PeterA: One alternative to this; could be to create the zip using some other background process (independent of the web server).
Whilst it creates the zip send the content.  If a user "pauses" (or drops or something), the zip is still created, just the chunks are no longer sent to the client.

If a client re-requests a download with a range request, simply continue the request from the background zipping process.

All of this presupposes that the zip file is basically required.  Obviously generating the full zip file if you don't need it is (almost? - context dictates if it is almost!) as wasteful as generating the start of a zip file and discarding the parts up until the start of the range request.

http://blog.stephencleary.com/2016/11/streaming-zip-on-aspnet-core.html#comment-8db8ed10-9488-11e8-a1dd-499116808ad6