(Staticman) Amir: Hi Stephen, Great Blog!!
After reading your tour of tasks section and this post, then I get to the following conclusions:

1. Delegate tasks must run on some kind of Thread-pool thread since you must use a thread to run CPU bound code.
2. Promise tasks are actually the real async operations. They use a Task an a "interface" wrapper but no thread is allocated since there is no use for one.

Am I correct here? 

Tasks are complex and I couldn't find an in-depth blog (yours is the closest). Most of the blogs focus on the async state machine. Can you also recommend a source for a behind the scenes on Tasks? I'm currently reading the source code using ILSpy and reading your blog, but I would like to go deeper.

https://blog.stephencleary.com/2013/11/there-is-no-thread.html#comment-178e2d80-9ad7-11e8-82d8-69ed00e8bcf5