(Staticman) : Hi Stephen,
I am running a task using Task.Run. This task is kind of a background task that just waits asynchronously for items to be added to the queue(Calls ReceiveAsync on BufferBlock class), and then processes the item. Since this task is an ongoing task it has to keep running and I do not await it.

All of this is part of class library which can be used by any client (console app, win forms etc)

I want any exception raised within the task to be rethrown to the client. However, since the task is not awaited, exceptions are swallowed.

Is there any way around this problem? 

Thanks

https://blog.stephencleary.com/2013/11/taskrun-etiquette-examples-dont-use.html#comment-89b00f50-ab8f-11e8-b6e7-61e1106727d7