(Staticman) Stephen Cleary: Right now, the client code does not get a signal that the processing has completed. It just shoves it in a queue and the processing task takes care of it. In order to notify the client of exceptions, you'll need to pass a signal back to the client.
One way to do this is to modify your queueing method so that it creates a `TaskCompletionSource<object>` and returns its `Task` to the client. This gives the client a "handle" to know when that item has completed processing. You would queue this `TaskCompletionSource<object>` along with the item to the queue, and the processing task should attempt processing and then complete the `TaskCompletionSource<object>` either successfully (`TrySetResult`) or with a processing exception (`TrySetException`).

Alternatively, if you want *one* task for your client to observe exceptions on for all items, you can just expose the `Task` that is returned from `Task.Run`.

https://blog.stephencleary.com/2013/11/taskrun-etiquette-examples-dont-use.html#comment-8d0d28f0-abcf-11e8-b6e7-61e1106727d7