(Staticman) Stephen Cleary: In this case, the first thing to do is send a request to the library maintainers for asynchronous APIs. :)
In the meantime, what *you* have to deal with is a synchronous, blocking API. Wrapping it with `Task.Run` would create what I call a "fake asynchronous" method. In an ASP.NET context, you should not use `Task.Run` to wrap it. That would just cause a thread switch for no benefit.

On the other hand, if you were on a UI thread in a client app (e.g., WPF / Xamarin / Windows Forms), then using `Task.Run` to wrap it would be acceptable - it would give you the benefit of freeing up the UI thread.

https://blog.stephencleary.com/2013/11/taskrun-etiquette-examples-using.html#comment-3f4408a0-b10b-11e8-8dc2-1511f82e59f7