(Staticman) Steven Bone: Hi Stephen,
Recipe 2.3 in your (most excellent) book is not as extensive as this post!  You specified the problem in the post and the above comment regarding how a progress method can come in after the final task.result and hinted `We’ll cover more advanced progress composition in a later post, and consider solutions to these problems.` Can you provide a link to this discussion or further advice? I apparently missed it and don't grasp an immediate solution.

https://blog.stephencleary.com/2012/02/reporting-progress-from-async-tasks.html#comment-9c3f3860-b4fb-11e8-a1b7-df997e15fc27