(Staticman) Steve: Thanks for this really useful post!
Things start to get tricky if you try to make a thread-safe version of this class. Since the arguments to CollectionChanged are only meaningful relative to the state of the collection at the time the collection was changed, we cannot allow any intervening operations between the actual change to the collection and the CollectionChanged event callbacks.  This means that the CollectionChanged event has to be fired from within a lock -- clearly a recipie for deadlock if one of the event listeners happens to invoke a method in the collection. 

What I ended up doing was allowing re-entrancy for the purposes of reading the collection, but for a write operation I timeout the lock and throw an exception saying that deadlock (likely) occurred. This is somewhat analogous to the exception thrown by, for example, List if you try to modify the List while iterating over it. 

To do it, one requires two locks and a private List. In pseudo-code (actually real code but I'm skipping the details) it looks a bit like the following. Here "ExcludeOtherWriters" and "ExcludeOtherReaders" are SemaphorSlim-based locks.  

        public void Insert(int i, T item)
        {
            using (this.ExcludeOtherWriters.WaitOrThrow(this.DeadLockDetectionMilliseconds)) {
                NotifyCollectionChangedEventArgs added;
                using (this.ExcludeOtherReaders.Wait()) {
                    this.List.Insert(i, item);
                    added = new NotifyCollectionChangedEventArgs(NotifyCollectionChangedAction.Add, item, i);
                }
                this.CollectionChanged?.Invoke(this, added); // synchronous
            }
            base.NotifyPropertyChanged(nameof(this.Count));
            base.NotifyAllIndexedPropertiesChanged();
        }

sjb-sjb

http://blog.stephencleary.com/2009/07/interpreting-notifycollectionchangedeve.html#comment-86cf99a0-d1b8-11e8-94ce-b56b73f5fd7f