(Staticman) Stephen Cleary: 1)
The statement `await Moe();` is the same as saying `Task t = Moe(); await t;`. In words, what is happening is that we are calling the method which returns a task, and then we await the task.

There are several possible scenarios where a task may have completed before the `await`. It may represent a cache hit; if `Moe` downloads a resource that is already in an in-memory cache, then `Moe` may return a task that is already completed. Or it could just be a very fast task; if `Moe` just delays for 1ms and the calling thread is swapped out at just the right time, then `Moe` could return an incomplete task that is completed before the `await` is reached.

2)

Yes, the "remainder" is Part B. That is, it continues executing the next statement after the `await`.

"returns from the `async` method" is referring to returning from the method that has the `await`; in the original example, this would be returning from `DoSomethingAsync`.

https://blog.stephencleary.com/2012/02/async-and-await.html#comment-fac3ef20-d607-11e8-9cf8-c9b6ddd6c0e8