(Staticman) Stephen Cleary: I'm not a "team"; I'm just one person.
I took a quick look at the code you posted, but it's rather overwhelming. I did see that it is doing some unzipping/deflating, which will not run asynchronously. If by "unresponsive" you mean "blocks the calling thread for a bit when downloading something huge", then that could be the case.

I recommend reducing the code in your question to a [minimal, complete, and verifiable](https://stackoverflow.com/help/mcve) amount.

https://blog.stephencleary.com/2012/02/async-and-await.html#comment-c170e190-d7d0-11e8-a9f1-05b1ca7b5c0d