(Staticman) Stephen Cleary: If you want to express cancellation, I think the best option would be to cancel the task.
So, if the dialog closes normally, you would use `TrySetResult`, and if the user cancels the dialog, you would use `TrySetCanceled`.

In general, I do not try to use `async`/`await` to control "user flow" like this in an application. I do not use "awaitable dialogs". In my experience, it works for simple things but quickly becomes too complex if the user has multiple choices. A more realistic (but also more complex) pattern for user interaction is Reactive Extensions (Rx). There's a fascinating project called [ReactiveUI](https://reactiveui.net/) that is specifically for defining your UI using Rx-enabled components, which makes a lot of sense to me (though I've never actually used it).

https://blog.stephencleary.com/2012/02/creating-tasks.html#comment-5402faf0-d92c-11e8-aeb2-a14cbc5b8eb2