(Staticman) Stephen Cleary: Your code needs the context whenever it depends on something the context provides.
That's a horrible answer, I know. :)

But the answer really does depend on which context you're talking about and what the code is doing. Here's the most common examples:
- UI components must be accessed by their owning thread. The UI context will ensure that the code after `await` will resume running on the UI thread.
- `HttpContext.Current` should only be accessed from within a request context. The ASP.NET Classic context will ensure that the code after `await` will be in the same request context. (ASP.NET Core does not have `HttpContext.Current` and does not need a context).

More generally, though, it's possible for you to create your own hosting system that has some definition of a "context", and you can write a `SynchronizationContext` that will integrate with `async`/`await` to resume on that context after every `await`. I've done this with my `AsyncContext` class, which is a thread with a context that always resumes on that thread, so it's like the UI context but without any UI capabilities (it's not an STA thread, does no message pumping, etc).

https://blog.stephencleary.com/2012/02/async-and-await.html#comment-3c4f7920-1355-11e9-bddf-5fc5f93eb3d7