(Staticman) Stephen Cleary: I suspect that this is due to this:
> The code uses tasks, but no async/await.

For the implicit async context to actually work, at some point you have to establish a copy-on-write scope. In the `async` world, this is handled by the `async` keyword itself; the compiler-generated state machine establishes a copy-on-write scope at the beginning of every `async` method. I *expect* that xUnit's test runner does not call your code from an `async` method, and since all your tests are synchronous, they would all be sharing the same scope. That is, assuming your `AsyncLocal<ConcurrentDictionary<..>>` is a static instance, which seems likely.

One way to test this is to make your unit tests `async` (you don't have to use `await`), and see if that clears up the problem.

If that doesn't clear it up, try to create a minimal repro and open an issue with the xUnit team (or post on Stack Overflow) and send me a link to it.

https://blog.stephencleary.com/2013/04/implicit-async-context-asynclocal.html#comment-61f24610-152d-11e9-82d7-0bb39bf6d207