(Staticman) Stephen Cleary: There's no way to say for sure, because if you block on `async` code, then you are depending on the implementation of that code to prevent deadlocks.
`HttpClient` generally strives to allow blocking code. On most platforms - I have seen it deadlock on mobile platforms. And of course you're taking a dependency on the *implementation* of `HttpClient` - it is entirely possible that in the future a bug may be introduced. So even if it works today, it's not the kind of thing I'd want in my codebase.

But why are you asking about blocking on `HttpClient` methods? It's weird to call asynchronous methods and then block on them. Why not just use `WebClient` if you want to block?

https://blog.stephencleary.com/2012/07/dont-block-on-async-code.html#comment-f5c804f0-19a7-11e9-a165-97c27510dd3f