(Staticman) Stephen Cleary: Hi Mike,
Thanks for your comments. They illustrate why it is difficult for me to explain what I mean when I say "asynchronous". In particular, I wrote this blog post specifically to go into detail (on one specific platform) explaining how it's possible to be asynchronous without threads (or processes, or whatever). This is one of the biggest questions people have after reading async intro posts [like this one of mine](https://blog.stephencleary.com/2012/02/async-and-await.html).

"Asynchronous" (as used in the modern sense) does *not* mean "runs on a different thread" (or process, or whatever). The modern meaning of "asynchronous" is what us old-timers would refer to as "callbacks" - just a lot easier to use these days. It's true that asynchronous *used* to be more synonymous with "parallelism" or "multithreading", but its meaning has shifted over the last decade or so.

https://blog.stephencleary.com/2013/11/there-is-no-thread.html#comment-0bd991e0-1b5c-11e9-ba78-3d3ffff9172b