(Staticman) Felix: This code works nice. In my use case a single request is rather slow; I need to send batches of requests in parallel and stream them into the zip in order to gain more speed. Here it gets nasty.
When I just do it using a semaphore, I get `Entries cannot be created while previously created entries are still open`, which kind of make sense. When I use your `LockAsync`, it turns out the same as running the requests in sequence.

I was thinking about a queue with event handlers, but I did not find much about it. And I don't know how to get the zipping code into an Enqueue event handler.

Do you have an idea how to achieve this in a simple way?

https://blog.stephencleary.com/2016/11/streaming-zip-on-aspnet-core.html#comment-e473ae00-1c3d-11e9-85a5-7f61d3b02144