(Staticman) astrowalker: "Most people would not write code like this. It’s very unnatural to call Task.Wait in an async method"
Exactly like this of course not, but at some point you will have to wait this way or another. For example in console program you won't have any explicit `Wait` but the program will wait implicitly, causing a deadlock as well. So this is not some rare case,, all it takes that Try/Set methods are used in non-async block. Add `await Task.Delay()...` before the call to `Try/Set` call and everyting is fine.

http://blog.stephencleary.com/2012/12/dont-block-in-asynchronous-code.html#comment-c3677b50-22d8-11e9-897b-9f0ee46a84c2