(Staticman) Stephen Cleary: I stand by my statement that blocking on asynchronous code is unnatural.
In server-side applications like ASP.NET Core, the entire stack is asynchronous. There is never any blocking, even within the ASP.NET runtime or IIS.

Client-side GUI applications are also a natural fit for asynchrony; the UI thread runs a main loop, and async/await interacts with that loop natively. Again, there is no blocking necessary.

Console applications are the *one* place where blocking is normal, and it's only normal to block in one line of code in the entire application, in order to prevent the main thread from exiting. What's more, modern platforms allow `async Main`, which hides that blocking in the runtime.

https://blog.stephencleary.com/2012/12/dont-block-in-asynchronous-code.html#comment-c3a23690-2329-11e9-b607-f720b75a151c