(Staticman) Sanjeev: Hi Stephen,
I had the confusion around the UIContext and the UIThread. Since I have read somewhere that for the UIContext we have a single UIThread and hence it runs on the same thread before and after the asynchronous operation (when not using cofigureawait). Also I tried checking the threadId while debugging and got it same.

So then it means that the following proposition is correct in case of UIContext.

// This code is running on UIThreadX

await DoSomethingAsync();

// This code should continue on the same UIThreadX

If it is then it means that in case of UIContext, we have thread affinity.

Please clarify my doubt. 

And also can you give me some pointer for the ASP.NET request context as to how the Thread changes in the ASP.NET request context after an async operation.

https://blog.stephencleary.com/2012/02/async-and-await.html#comment-4a2f7c70-232c-11e9-b607-f720b75a151c