(Staticman) Stephen Cleary: Yes, the UI context is tied to a particular thread, and will resume on that same thread.
The ASP.NET context is different. When a request comes in, ASP.NET grabs one of its thread pool threads and has it handle that request. In the case of an asynchronous request, when your handler code yields back to the ASP.NET runtime, that thread pool thread is returned to the thread pool. It is no longer associated with that request. Later, when your handler code is ready to resume, ASP.NET will grab one of its thread pool threads and use it to continue the request. This may be the same thread pool thread or it may be a different one.

For more information on `async` in ASP.NET, see my [article on `async` ASP.NET](https://msdn.microsoft.com/en-us/magazine/dn802603.aspx).

https://blog.stephencleary.com/2012/02/async-and-await.html#comment-57b266f0-2331-11e9-b607-f720b75a151c