(Staticman) Stephen Cleary: In his article, he states:
> The temporary variable ensures that this is “thread safe”.

and then later states:

> You've got to solve both problems if you want to do threadsafe events.

I.e., the temporary variable is insufficient to make a threadsafe event.

Most of the confusion around this issue is around the question "what does threadsafe mean". And both of our articles agree that there is *no way* to make events threadsafe just on the producer side. Eric's second statement sums that up.

Eric mentions a solution you could make on the consumer side, but it's not easy to get right (or even to know you need to do it).

In my opinion, the core issue here is that C# events are rather poorly designed, and lead developers into a pit of failure as soon as multiple threads enter the mix. `IObservable<T>` is how events *should have been* designed; it provides a new producer *and* consumer pattern that is truly threadsafe. In my own code, I *only* use events if they are always raised from the same thread; that way, thread safety is trivial.

The only point in which Eric and I disagree is whether the temporary variable is useful. It is useful for avoiding null reference exceptions when using cross-thread events, but it is not sufficient to avoid race conditions when using cross-thread events (we both agree on all this). I always recommend not using cross-thread events *at all*, in which case the temporary variable is not useful. He recommends allowing cross-thread events, with the understanding that your event handler must have special logic to detect and handle race conditions.

https://blog.stephencleary.com/2009/06/threadsafe-events.html#comment-96a9aee0-23e0-11e9-aa77-ffe40b10080e