(Staticman) Felipe Longo: Hi Stephen. Awesome material.
About the subject of passing CancellationToken to Task.Run
Even if there is a lot of Tasks being created in a System (many different calls to Task.Run at the same time), it doesn't bring that much of a benefit? 

If a Task is not running yet the CancellationToken doesn't cancel it?  
Or what matters is the Status = Scheduled? 

Thanks in advance :D

http://blog.stephencleary.com/2015/03/a-tour-of-task-part-9-delegate-tasks.html#comment-306b0a40-2f19-11e9-9447-21eee9ff7847