(Staticman) Stephen Cleary: [I confirm this behavior](https://blog.stephencleary.com/2012/12/dont-block-in-asynchronous-code.html); the cause is an optimization in `await` where it schedules its continuations with the `ExecuteSynchronously` flag.
There isn't a great way around this; there's no way to tell `await` to not use that flag. The only real options are to not block in asynchronous code (which I do recommend!), and passing the `TaskCreationOptions.RunContinuationsAsynchronously` flag to the `TaskCompletionSource<T>`, which overrides the `ExecuteSynchronously` flag by forcing the continuation onto a new thread pool thread.

https://blog.stephencleary.com/2012/07/dont-block-on-async-code.html#comment-82ab86c0-3763-11e9-acc4-e918b93ed4b3