(Staticman) dbnex14: Thanks Stephen.  One thing though.  I think in your 2nd best solution (or go async all the way), you are misnaming the top method and lib method.
Button1_Click is calling GetJsonAsync, so the method Get inside MyController should be renamed to GetJsonAsync.  That method then calls GetStringAsync inside a using block like you did in the original example under UI Example at the top of the article.  

I have read this post 2-3 times so far and every time this part confuses me and I am not sure if it is correct.  Would it be possible to improve and correct this?  Or I am again missunderstanding your last point?  

I am just trying to match your solution to the problem at the top and they do not match in terms of method names and the missing using block.  Thanks again

https://blog.stephencleary.com/2012/07/dont-block-on-async-code.html#comment-5b752610-3ab0-11e9-b9e3-a34129e9fd77