(Staticman) Stephen Cleary: I couldn't say. The code looks correct; usually this kind of error happens if you forget a `using` statement, but all of yours appear to be present.
Are you accessing the file before it has finished downloading? One of the issues with the "compress on the fly" approach is that the response headers do not have any content length information, so the browser doesn't know how much of the transfer is remaining until the transfer is done.

If that isn't the problem, try capturing the data in Fiddler when running each program and seeing what the difference is. Or it could be a Windows client problem: antivirus or "blocked" downloads can interfere with opening downloaded files. Try opening the file with a third-party application like 7-Zip.

https://blog.stephencleary.com/2016/11/streaming-zip-on-aspnet-core.html#comment-cb7c0f70-3bb7-11e9-97f7-73cce500929f