(Staticman) Stephen Cleary: Putting data, business, and view logic in one file is an extreme misuse of React. And that wouldn't be using Redux at all.
With Redux, business logic always goes in your reducers, and can be in one or multiple files. Likewise, data logic would normally go into a separate file, to keep the reducers as simple as possible.

The point of React is to keep your view logic (for a single component) in a single file, instead of spread over multiple files.

https://blog.stephencleary.com/2016/02/why-react-why-redux.html#comment-ac972240-4191-11e9-9e1e-cb0bd9d7f61d