(Staticman) Sanjeev: Hi Stephen,
In UWP/Rich Client Application: The way an async method resumes on the UI thread is through synchronizationcontext. 
For ex:

async Task DoWorkAsync()
{
    await asyncOp(); // this method may be making a network call
    continuation
}

So it means when the Task(represented by asyncOp()) gets completed, then a thread in thread pool thread typically an iocompletion thread uses the synchronizationcontext's post method to send the continuation/callback to message queue so that the UI thread can pick it up using the message loop and hence the continuation runs back on the UI thread.

Please correct my understanding if I am wrong.

Also in case of await Task.Delay(1000); how do we resume back on UI thread? Since in network async call we have iocompletion port but in Task.Delay which uses underlying Timer class, which is the thread which makes a synchronization.post() call?

http://blog.stephencleary.com/2012/02/async-and-await.html#comment-f1318270-4681-11e9-8eaa-bf9870a70db5