(Staticman) Stephen Cleary: I'm not sure if I fully understand what you're asking, but I'll do my best to answer.
I/O work is generally passed to another system. E.g., with DMA, the device can access the memory directly. There's no CPU involved, and thus no thread at all.

To *complete* the I/O work, some notification does need to come in. This kind of notification (eventually) usually ends up on a thread pool thread, which is what actually transitions the `Task` object to the completed state. This in turn queues up the continuation of the `Log` method.

https://blog.stephencleary.com/2012/02/async-and-await.html#comment-309fc630-7051-11e9-936a-099b6a65bff3