(Staticman) Stephen Cleary: > Is there any way to check from code that there is no context?
Yes, but it's awkward. There is no context when there is no `SynchronizationContext` and no `TaskScheduler`. There is no `SynchronizationContext` when `SynchronizationContext.Current` is `null` *or* it is an instance of the `SynchronizationContext` class and not a derived type. There is no `TaskScheduler` when `TaskScheduler.Current` is not the same as `TaskScheduler.Default`.

A better question to ask is: why would my code care whether there is a context or not? Can I change it so that it doesn't care?

> on WPF it is set, on .NET Core console it is null but on Full framework console it is also null

Every UI app will have a context for its UI thread. This is true across WPF/WinForms/Xamarin/etc. Console apps do not start with a context on their main thread, regardless of framework.

> when I run some awaited method with ConfigureAwait(true) when context is null I’m returning to the same thread

This is possible, but not guaranteed. If your code is running on a thread pool thread, then resuming without a context will schedule the continuation to a thread pool thread, which may or may not be the same thread pool thread that started the method.

> can I execute some code to check if I have context and if I really don’t have context like on ASP.NET can I return somehow to original thread?

Contexts don't guarantee returning to the same thread. ASP.NET pre-Core is an example of that: it provides a custom `SynchronizationContext` that can use *any* thread pool thread. Custom task schedulers are another example; e.g., `ConcurrentExclusiveSchedulerPair` is a context that can use any thread pool thread.

The vast majority of code doesn't *need* to return to the same thread. If your code does, then it should manage its own context. I have an [`AsyncContext` type](https://github.com/StephenCleary/AsyncEx) that enables this, but before you use it, make *absolutely sure* that you *need* it.

https://blog.stephencleary.com/2017/03/aspnetcore-synchronization-context.html#comment-fd5338f0-8213-11e9-9106-19ed2574ea23