(Staticman) sjb: I keep reading this post over and over, it is so useful!
There is a real 'gotcha' with the NotifyCollectionChangedEventArgs constructors. If you are working with a typed collection then you may well have your newItems and oldItems typed as IList<T> rather than IList. If you pass these into the constructor then you will get zapped with a rather unobvious exception, such as 'attempt to access beyond memory' or similar. Since IList<T> does not implement IList, the constructor that is actually used is the one with 'object' as the element type, not the one that takes IList. In other words you end up trying to notify a single-element change with the single element being an IList<T>. 

I made this same mistake twice, which to me is a sign that a helper method is needed.

https://blog.stephencleary.com/2009/07/interpreting-notifycollectionchangedeve.html#comment-3a61fab0-8548-11e9-ba28-3542af84de90