(Staticman) Stephen Cleary: > Are there any issues here with cold start time of Azure Function (on the consumption plan at least)?
Yes! Serving the website using this technique does make it subject to the regular AF cold start behavior. There is a new Premium Azure Functions offering announced a month or two ago, where you can keep one pre-warmed instance to avoid cold start. Of course, the downside is you have to pay for it... :)

> Are there any reasons to use this approach over "Static website hosting in Azure Storage"

In my case, I wanted my API and website to use the same domain. You should be able to get a static website with an API, but you'd need different domains and set up the CORS correctly.

https://blog.stephencleary.com/2017/08/azure-functions-spa.html#comment-661ac4a0-9904-11e9-836a-73a1def69d80