(Staticman) Morgan: Thanks for the article and a lot of your blog posts about async/await, they've definitely helped me learn more about it.
I noticed that your search for SynchronizationContext was in the aspnet/Mvc repo. I performed the same search on the dotnet/corefx and aspnet/aspnetcore repos and found results for SynchronizationContext.

https://github.com/dotnet/corefx/search?q=SynchronizationContext&unscoped_q=SynchronizationContext
https://github.com/aspnet/AspNetCore/search?q=SynchronizationContext&unscoped_q=SynchronizationContext

With the corefx repo containing a definition for SynchronizationContext.

https://github.com/dotnet/corefx/blob/master/src/Common/src/CoreLib/System/Threading/SynchronizationContext.cs

To me it seems like dotnet core and aspnet core contain a SynchronizationContext (based on their repos) but would you be able to help clarify my understanding of this?

https://blog.stephencleary.com/2017/03/aspnetcore-synchronization-context.html#comment-339a18c0-d0ac-11e9-9268-fba15f9b56fc