(Staticman) Stephen Cleary: The base `SynchronizationContext` type in the `corefx` repo is a context that represents the thread pool context. By convention, a `SynchronizationContext.Current` of `null` is semantically equivalent to a `new SynchronizationContext()`. More info: https://msdn.microsoft.com/en-us/magazine/gg598924.aspx
As Dave mentioned, the `RendererSynchronizationContext` in the `AspNetCore` repo appears to be a synchronization context used by Blazor. I am not very familiar with Blazor, but I assume this context is used when running rendering code on the server side, and possibly may exist on the client side (in the browser) as well.

https://blog.stephencleary.com/2017/03/aspnetcore-synchronization-context.html#comment-96345c90-d410-11e9-985e-d7733b393af6