(Staticman) Justin Swanson: Hey Stephen!  Thanks for your continued work on educating everyone on all these amazing concepts. 8)
You say explicitly "do not use Task.Run in the implementation of the method; instead, use Task.Run to call the method."
That makes sense to me, for most heavy things, let the user decide.  This seems to follow the advice I've heard a lot of other places:  never use Task.Run() in a library.

My question, though, is how to do parallel CPU-bound work that is supposed to do the work on multiple cores using the clean API tasks provide.  You suggested in an earlier post that perhaps you use Parallel.ForEach or something and still offer it in a synchronous method the user can choose to Task.Run() on.  However, the Parallel library is not the cleanest API to use.  I'd much rather use the API offered by tasks, where I can add multiple tasks to a list of work to be done, and then Task.WhenAll which returns me their results in the desired order.  Is there really no place for multicore threading that makes use of the Task pattern, where chunks of work are offloaded to threads via Task.Run() and then collected via WaitAll?  I understand that in an ASP server application, you need to be aware of the resources you're using, so you want to limit concurrency via a parallel max concurrency parameter.  However, for an application that simply wants to calculate something as fast as possible and burn as many cores as possible, is Task.Run() + WhenAll still not a proper solution?

Thanks much in advance!

https://blog.stephencleary.com/2013/11/taskrun-etiquette-examples-dont-use.html#comment-5c058420-d50c-11e9-ab7b-33d7aa1519e2