(Staticman) Stephen Cleary: First, the guideline against `Task.Run` is primarily to prevent "fake asynchrony" - methods that *appear* asynchronous but actually just run synchronously on a background thread. This was a common pitfall for libraries at the time this was written; today, it's generally understood that this is a bad practice.
If your library is specifically about CPU-bound processing that can be parallelized, then it is appropriate to use `Parallel` (or `Task.Run`) in that library. You just need to clearly document it. If possible, I would structure the library's API so that it wouldn't do its own parallelism, leaving that up to the consumer. But that is not always possible.

For CPU-bound work, I find `Parallel` or PLINQ to be more convenient than `Task` APIs. They are at a higher level and easier to use IMO (once you find the right one(s)).

https://blog.stephencleary.com/2013/11/taskrun-etiquette-examples-dont-use.html#comment-75413eb0-d59e-11e9-a060-892b2dcbd404