(Staticman) Stephen Cleary: The code sample there is unusual; it was specifically written to demonstrate the issue. It's very rare to have code on ASP.NET that modifies shared data like that.
So the best way to avoid it is just not to write code like this in the first place. Asynchronous code in particular flows much better in a functional style: *returning* results rather than updating some shared data location.

That said, if you do find yourself in a situation where it's not easy to change code like this, an async-compatible synchronization primitive could be used to coordinate access, e.g., `SemaphoreSlim` with its `WaitAsync` and `Release` methods.

https://blog.stephencleary.com/2017/03/aspnetcore-synchronization-context.html#comment-c0e7fce0-d644-11e9-95ba-c123820aeaaa